seperated entities configuration page from settings
