+++ src/App.js
@@ -6,6 +6,7 @@ import React, { useState, useEffect } from "react";
 import { io } from "socket.io-client";
 import { HashRouter as Router, Route, Routes } from "react-router-dom";
 import SplitLayout from "./views/layouts/SplitLayout";
+import Entities from "./views/Entities";
 
 function App() {
   const [socketInstance, setSocketInstance] = useState("");
@@ -61,6 +62,7 @@ function App() {
               element={<SplitLayout socket={socketInstance} />}
             />
             <Route exact path="/settings" element={<Settings />} />
+            <Route exact path="/entities" element={<Entities />} />
             <Route path="*" element={"not found"} />
           </Routes>
         </Router>
