+++ src/views/Settings.js
@@ -1,18 +1,12 @@
 import ToolBar from "../components/ToolBar";
-import EntityConfig from "../components/EntityConfig";
-import { v4 as uuid } from "uuid";
-import { useNavigate } from "react-router-dom";
 import {
   Container,
   Form,
-  Button,
   Row,
   Col,
-  Spinner,
   InputGroup,
-  Alert,
 } from "react-bootstrap";
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import "../static/css/settings.css";
 
 export const defaultSettings = {
@@ -27,148 +21,17 @@ export const defaultSettings = {
 };
 
 function Settings() {
-  const [current_entities, setCurrentEntities] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [showError, setShowError] = useState(false);
-  const [errorMsg, setErrorMsg] = useState("");
-  const navigate = useNavigate();
 
   useEffect(() => {
     getCurrentSettings();
-    getCurrentEntities();
   }, []);
 
-  const getCurrentEntities = () => {
-    fetch("http://localhost:5000/entities_config")
-      .then((res) => res.json())
-      .then((res) => {
-        setCurrentEntities(res);
-        setLoading(false);
-        validateCurrentConfig(res);
-      });
-  };
-
-  const addEntity = () => {
-    fetch("http://localhost:5000/default_entity")
-      .then((res) => res.json())
-      .then((res) => {
-        // res.id = parseInt(Math.random() * 10000);
-        res.id = uuid();
-        setCurrentEntities([...current_entities, res]);
-      });
-  };
-
-  const removeEntity = (id) => {
-    const remaining = current_entities.filter((e) => e.id != id);
-    setCurrentEntities(remaining);
-  };
-
-  const getValue = (id) => {
-    return document.getElementById(id).value;
-  };
-
-  const getDropdownValue = (id) => {
-    return document.getElementById(id).textContent;
-  };
-
-  const getEntityConfig = (eId) => {
-    const _useXYZ = document.getElementById(`useXYZ-${eId}`).checked;
-    const _useRPY = document.getElementById(`useRPY-${eId}`).checked;
-    const wireframe = document.getElementById(`wireframe-${eId}`).checked;
-    const tracked = document.getElementById(`tracked-${eId}`).checked;
-
-    const position = _useXYZ
-      ? {
-        table: getDropdownValue(`positionTable-${eId}`),
-        x: getDropdownValue(`x-${eId}`),
-        y: getDropdownValue(`y-${eId}`),
-        z: getDropdownValue(`z-${eId}`),
-      }
-      : {
-        table: getDropdownValue(`positionTable-${eId}`),
-        longitude: getDropdownValue(`lon-${eId}`),
-        lattitude: getDropdownValue(`lat-${eId}`),
-        altitude: getDropdownValue(`alt-${eId}`),
-      };
-    const attitude = _useRPY
-      ? {
-        table: getDropdownValue(`attitudeTable-${eId}`),
-        roll: getDropdownValue(`roll-${eId}`),
-        pitch: getDropdownValue(`pitch-${eId}`),
-        yaw: getDropdownValue(`yaw-${eId}`),
-      }
-      : {
-        table: getDropdownValue(`attitudeTable-${eId}`),
-        q0: getDropdownValue(`qw-${eId}`),
-        q1: getDropdownValue(`qx-${eId}`),
-        q2: getDropdownValue(`qy-${eId}`),
-        q3: getDropdownValue(`qz-${eId}`),
-      };
-
-    const config = {
-      name: getValue(`name-${eId}`),
-      alpha: parseFloat(getValue(`alpha-${eId}`)),
-      useRPY: _useRPY,
-      useXYZ: _useXYZ,
-      pathColor: getValue(`pathColor-${eId}`),
-      wireframe: wireframe,
-      color: getValue(`color-${eId}`),
-      tracked: tracked,
-      scale: parseFloat(getValue(`scale-${eId}`)),
-      position: position,
-      attitude: attitude,
-    };
-    return config;
-  };
 
   const parseLocalStorage = (key) => {
     var value = localStorage.getItem(key);
     if (value === "" || value === null) value = defaultSettings;
     else value = JSON.parse(value);
     return value;
-  };
-
-  const applyConfig = async () => {
-    // Entity Configs
-    const configs = [];
-    current_entities.forEach((e) => {
-      const c = getEntityConfig(e.id);
-      configs.push(c);
-    });
-
-    const response = await fetch("http://localhost:5000/write_config", {
-      headers: {
-        "Content-Type": "application/json",
-      },
-      method: "POST",
-      body: JSON.stringify(configs),
-    }).then((res) => res.json());
-
-    if (!response.ok) {
-      setErrorMsg(response.msg);
-      setShowError(true);
-      window.scrollTo(0, 0);
-    } else {
-      setErrorMsg(response.msg);
-      setShowError(false);
-      navigate("/home");
-    }
-  };
-
-  const validateCurrentConfig = async (entities) => {
-    const response = await fetch("http://localhost:5000/validate_config", {
-      headers: {
-        "Content-Type": "application/json",
-      },
-      method: "POST",
-      body: JSON.stringify(entities),
-    }).then((res) => res.json());
-
-    if (!response.ok) {
-      setErrorMsg(response.msg);
-      setShowError(true);
-      window.scrollTo(0, 0);
-    }
   };
 
   const getCurrentSettings = () => {
@@ -190,29 +53,11 @@ function Settings() {
     localStorage.setItem("general_settings", JSON.stringify(general_settings));
   };
 
-  const showSettings = loading ? "hide" : "show";
-  const showLoading = loading ? "show" : "hide";
-
-  function AlertError() {
-    if (showError) {
-      return (
-        <Alert variant="danger" onClose={() => setShowError(false)} dismissible>
-          <Alert.Heading>Invalid Config!</Alert.Heading>
-          <p>{errorMsg}</p>
-        </Alert>
-      );
-    }
-  }
-
   return (
     <>
       <ToolBar page="settings" />
-      <Container className={"loading " + showLoading}>
-        <Spinner variant="primary" />
-      </Container>
 
-      <Container className={"settings-page " + showSettings}>
-        <AlertError />
+      <Container className="settings-page">
         <Form>
           <fieldset>
             <legend>• View Helpers 🌎</legend>
@@ -296,40 +141,6 @@ function Settings() {
               </Col>
             </Row>
           </fieldset>
-          {current_entities.map((e) => (
-            <EntityConfig
-              key={e.id}
-              eId={e.id}
-              name={e.name}
-              pathColor={e.pathColor}
-              color={e.color}
-              wireframe={e.wireframe}
-              alpha={e.alpha}
-              useRPY={e.useRPY}
-              useXYZ={e.useXYZ}
-              tracked={e.tracked}
-              scale={e.scale}
-              position={e.position}
-              attitude={e.attitude}
-              removeEntity={removeEntity}
-            />
-          ))}
-
-          <Row>
-            <Col className="text-start"></Col>
-
-            <Col className="text-center">
-              <Button variant="success" onClick={addEntity}>
-                Add Entity
-              </Button>
-            </Col>
-
-            <Col className="text-end">
-              <Button variant="primary" onClick={applyConfig} type="submit">
-                Apply
-              </Button>
-            </Col>
-          </Row>
         </Form>
       </Container>
     </>
